X00TD: Do not enable updatable APEXes on FDE target
It can't really be enabled for this device, as apexd requires /data/apex
to be available right after boot to update all the Android modules.
As in all FDE devices, /data/apex is encrypted until the device
is unlocked by the user, rendering APEX basically useless as only
the system APEX variants will be loaded at boot.

Change-Id: I0649862e5c5b2c1a1cb597914a11df79b3e01957
Signed-off-by: kubersharma001 <kubersharma001@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>